used once to make payload of recursion semidet and recursion TCO
I was struggling to understand why my trivially tail call written
predicates are not being optimized until today. I have learned that if
they are non-deterministic the choice points are stored in the stack
and hence the stack is being blown. I used once/1 to make my recursion
payload semidet and then recursion was optimized.